build: correct disk image file creation, it was one byte too large
"dd if=/dev/zero bs=1 count=1 of=<diskimage> seek=<sizeinMB>*1024*1024"
resulted in a file that is desired size + 1 byte.  Subtract one byte
from the seek arithmetic, so that the write of 1 byte makes the disk
disk image exactly the right size. In the future it might be preferable
to use "qemu-image create -f raw <diskimage>" to do this operation.

Signed-off-by: James Perkins <james.perkins@linuxfoundation.org>
Signed-off-by: Jan-Simon Möller <jsmoeller@linuxfoundation.org>