Do not specify compression format for Arch packages when extracting
Arch packages are now using xz compression for packages, which were using gzip
compression before. When using GNU tar instead of bsdtar, they will be extracted
as gzip format instead, because the script explicitly tells tar that the package
uses gzip compression.

Remove the option for this and let tar guess the compression format of the
package, so both gzip and xz compressed packages could be handled.