changelog2spec: further refinement for date match
If a changelog entry header contains additional sequences of 4+ digits
immediately followed by an '@' character, all after the email address,
then the date match in this regexp will not work as intended.

For Tizen, changelog entry headers often have the form:

 * Fri Jun 07 2013 First Last <first.last@example.com> tagname@commitid

(The "tagname" above ends with a detailed timestamp with to-the-second
resolution, for example, 20130811.132529)

To fix these changelog entries, the "zero or more non-hyphen" match can
be made non-greedy to ensure that the 4-digit date is matched instead of
the last four digits from the timestamped tag following the email
address.