add OBS release number to Debian package version
The OBS auto-increment the RPM release number on every new build.
Debian has no direct support for release numbers,
but commonly used is to add the release to the version number, separated by "-":
<version>-<release>

Currently the Debian build process from OBS set the version to the string defined in the DSC file.
(A normal build will just use the version from the top most entry from debian/changelog.)

This patch against the build package adds support for auto-incremental release numbers also for DEB packages.

To use it, the project must provide a separat "debian.changelog" file
with at least one changelog entry and the DSC file must include the tag:
DEBTRANSFORM-RELEASE: 1

Therefore, the current build behaviour is not changed.
This new feature gets only activated, when the DEBTRANSFORM-RELEASE is added to the DSC file.