Support for simple images
Simple image is a really simple way how to get rootfs tarball or
squashfs image. The main advantage over existing solution like kiwi is
that it is distribution independent and therefore any distribution
supported by obs can make use of it. Disadvantage is that it is simple
and basically just preconfigured tarball - no bootloader, no kernel.

Syntax is pretty simple, all you need is to create simpleimage file and
write what you want inside of it. Here is the simple example, all fields
are optional as is the %build section.

            +----------------------------------+
            | Name:           example          |
            | Version:        1.0              |
            | BuildRequire:   emacs            |
            | #!BuildIgnore:  gcc-c++          |
            |                                  |
            | %build                           |
            | passwd << EOF                    |
            | opensuse                         |
            | opensuse                         |
            | EOF                              |
            +----------------------------------+