Revert "don't display 'keeping ..' messages. That's rather uninteresting."
This reverts commit e2a3a2681057293a452938cb523783f8d88809e6.

That commit crippled progress reporting. Issue 1: It made the
"keeping" messages go away, but the progress indicator would still be
printed, so the commit only did half of the intent.

Issue 2: Since no newline was emitted (this was part of the "keeping"
message) and because stdout is line-buffered, the progress report
would only be sporadically emitted.

Furthermore, initializating a Debian root is slow compared to RPM, and
so there is a long time before any output is shown.

Show me my progress as it happens. Reinstate messages about kept
packages.