allow skipping of unknown rpm specfile tags
Non-SUSE distributions (and maybe old SUSE distributions) do not know
RPM tags "Suggests", "Recommends", "Supplements".
This leads to ugly hacks needed in spec files if cross-distro build is
needed.
Test if these tags are actually supported by the target's RPM and just
drop them if not. This needs to be enabled explicitly in prjconf with
"Buildflags: dropunknownrpmtags"