Add support for building appliances with kiwi v8
kiwi version v8 is based on python and a rewrite of the former
perl based kiwi version. The design and caller semantic has
changed however kiwi v8 provides a compat call mode to allow
scripts to be adapted in a clean way. This commit adds a
is_legacy_kiwi method to check which kiwi is installed and
handle the call appropriately in perform_image_build|bundle