livebuild: parse /etc/debian_version on testing
When ran on Debian testing/unstable, lsb_release --codename will
return "n/a", which results in a broken config.
As a workaround, if "n/a" is returned, parse /etc/debian_version
which will contain "<testing codename>/sid".
The bug on lsb-release as been marked as wontfix:
https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=845651
That is fine for the distribution's use case, but for our use case
it makes sense to always return the testing codename.