Fixed use of provides tags
A provides tag of the format "foo:bar" causes a problem in rpm
if the %_wrong_version_format_terminate_build macro is not set
to 0. For suse builds the above macro is set to 0 and thus do
not cause problems. But for other distributions the build fails
if the kiwi packages creates provides tags in this format.

Therefore this patch is one out of series which changes the use
of the provides tags to the format "foo(bar)". There will be
reference patches to the spec files of the python-kiwi and
the kiwi-boot-descriptions package to complete the required
changes