computeblocklists: document and use alternate ioctl numbers
Commit 9ecb7dc7 added fallback ioctls for use on architectures that
had historical operating systems for which Linux reused ioctl
numbers.  The ioctl numbers were added as magic numbers, which makes
it difficult to determine what they are.

On these architectures, the ioctl direction field is 3 bits instead
of 2, and the "none" direction is represented as 1 instead of 0.  This
commits adds an alt_ioctl helper that takes the "standard" Linux
ioctl number and converts it to the ioctl numbers used on those arches.

We also now print consistent error messages between the two ioctl calls.