computeblocklists: add support for FIEMAP ioctl
The FIBMAP ioctl is deprecated and will do strange things under
certain circumstances (like with unwritten extents).  It's also
not supported at all by btrfs.  This commit adds support for the
FIEMAP ioctl and handles the flag to identify unwritten extents
(to be treated as holes) and other flags that mean the file range
can't be translated to a usable block range.  The FIEMAP ioctl
is much more efficient in the kernel, so we use that by default now.