computeblocklists: add fallback code if the "Q" type is not supported
Since the "Q" type is not available on all platforms or perl builds,
fallback code is needed in these cases. For this, instead of using
the unsigned quad type ("Q"), use two unsigned long (32-bit) types
("L").
To pack an unsigned quad, it is "halved" into two unsigned longs
(32-bit). Afterwards, the two unsigned longs are packed via "LL"
(taking the endianness into account (either little-endian or
big-endian (no "weird" byte orders))).
To unpack an unsigned quad, two unsigned longs (32-bit) are unpacked
and then represented by a Math::BigInt instance (again, taking the
endianness into account). Using a Math::BigInt potentially slows down
arithmetic operations.

This is a follow-up fix for commit 8c90847 ("computeblocklists: add
support for FIEMAP ioctl").

Note: it is not guaranteed that the "extractbuild" script can deal
with the results of the "computeblocklists" script. For instance,
assume:
- The perl that is used to execute the "extractbuild" script does not
  support 64-bit integers
- We have to read a block that cannot be represented by a 32-bit
  integer. Also, it cannot be represented as a perl floating point
  number (see "perldoc perlnumber").

Hence, most of the arithmetic that is performed with this block in
"extractbuild" results in "garbage". Using Math::BigInt might help
here, but then we would need to check if, for instance, a
seek(F, $some_big_int, 0) is translated to a _llseek/whatever...