build-recipe: Adjust check for email only authors to be specific for SLE
The way the condition was structured, setting %fullname_in_changelog to 1
would trigger the wrong behavior on non-SUSE systems.

This conditional evaluates to disable full identities only if it is a
SUSE system that isn't openSUSE only if %fullname_in_changelog is not set.

In all other cases, it preserves full identities if they exist.