debootstrap: Print the exact command we're going to run
debootstrap is not a simple command, and there are some subtleties to
the options we use, so we should log exactly what we did.

The incantation with `$(printf ' %q' "$@")` is to log the contents
of $@, correctly shell-quoted to avoid ambiguity (in case there are
spaces in an argument). It should expand to something like

    running debootstrap: chroot /buildroot debootstrap --etc --etc...

Signed-off-by: Simon McVittie <smcv@collabora.com>