build-recipe-dsc: Move all build results, not just *.deb and *.changes
The purpose of the changes file is to list everything that is to be
installed into the Debian (or derivative) archive, including *.deb,
*.udeb (micro-debs for the Debian installer), *.ddeb (detached debug
symbols, in Ubuntu and its derivatives) and *.buildinfo (details of the
build environment, mainly for reproducible builds). If we're providing
an equivalent of the Debian archive, then we should include all the
same files.

This change ensures that all build products get moved into the right
place, even if they are file types that obs-build doesn't yet know about,
or even file types that haven't been invented yet.

When obs-build is used in conjunction with OBS, changes to bs_publish
are also likely to be necessary for a full-stack solution.

Signed-off-by: Simon McVittie <smcv@collabora.com>