add ability to store ccache when PKG_CCACHE variable is true
This feature will allow usage of ccache previously created archive of
ccache within the build env.
The storage dir is /.ccache and permission for the $BUILD_USER are
already available.

The activation of feature is dependent on param `--pkg-ccache` being
passed. This parameter points to tar.gz archive. This archive will be
extracted to /.ccache

A .ccache.tar.gz is created in case of a successful build which can be
used for subsequent builds.