Autodetect whether to use --pipe option of systemd-nspawn.
The "--pipe" option of systemd-nspawn is useful for automated build,
but it leads to very inconvenient results when running interactive shell
using "osc shell" command: pressing Ctrl+C causes the whole session
to terminate.

Also, "--pipe" option is not available on older versions of systemd-nspawn,
like the one that is used in Raspbian 10.

This change makes build-vm-nspawn autodetect whether "--pipe" option
is available, and use it only for automated build, but not interactive shell.

Signed-off-by: Oleg Girko <ol@infoserver.lv>