Use grep -E instead of egrep to check for the needsbinariesforbuild flag
The reason is that egrep is implemented as a little shell script that
uses /bin/sh as interpreter. But /bin/sh is nowadays managed by the
alternatives mechansim, meaning it is not available until the
postinstall script of the shell is run. But postinstall scripts are
run in init_buildsystem, which is supposed to get the needsbinariesforbuild
option...