Fix vm_sysrq_kvm not working reliable
When connecting to the unix domain socket, qemu sends a greeting
message. If we just send the command and close the socket, the
linux kernel will see that there is unread data and reset the
socket, discarding our sent command. So we need to read the
greeting.

We do a shutdown before reading from the socket so that the qemu
side gets an EOF and closes the connection, terminating our read.