Add '--mode' option to build
Extend the '--stage' option:
Alongside the legacy form 'starge=-bSTAGE', which builds all rpmbuild stages
up to and including STAGE, support the forms:
 --stage=..STAGE to build everything up to STAGE,
 --stage=STAGE to build everything including STAGE,
 --stage=STAGE.. to build everything starting from STAGE.
The --stage=STAGE mode is equivalent to using  'rpmbuild --short-circuit'.
Of course, the previous stages must have been built and available in the
chroot environment. Minimal checks are performed to ensure this is the
case.
These options are useful to test changes to the %install or %files sections
of a spec file without having to go thru the entire %prep and %build stage.

Currently, this is available for RPM builds only.

Signed-off-by: Egbert Eich <eich@suse.com>