debtransform: keep epoch in 'Version' field of dsc file
Stripping the epoch from the version string and then writing it
back into the 'Version' field of the dsc file breaks debian packages.

By removing the epoch from the version string only once the updated
version string has been saved into the 'tags' map, it is kept in the
'Version' field of the dsc file.