add generic post build check script hook
Execute every script inside of /usr/lib/build/post-build-checks/ .

The check must validate that it runs for the right build type and fail
only when it finds a problem with a build result. No matter of which
type.

We already have usr/lib/build/checks/ but the existing checks are rpm
generic and fail on non rpm builds. Ideally these get migrated later on.

OBS-110