Force locale to C.UTF-8
Multiple packages have run into build failures on our OBS. On closer
inspection, it was concluded to be issues with the locale settings.

These issues were never seen on the Debian buildd infrastructure because
there the locale setting LC_ALL=C.UTF-8 is in use.

This change keeps us closer to what is used in Debian buildds.

https://github.com/openSUSE/obs-build/pull/665

Signed-off-by: Emanuele Aina <emanuele.aina@collabora.com>