Change KVM Console for s390x
IBM is recommending ttyS0 and ttyS1 (together) as types of kvm console in bbc#1187264. Before hvc0 (the same as for Power) has been used.
ttyS0 and ttyS1 have got the benefit that we can receive get kernel messages on both consoles.
ttyS0 is used as a default setting for all architectures (above of the file). I have changed hvc0 to ttyS1 for s390x.