turn off metadata_csm
not all older kernels support it, and it really doesn't matter
for build jobs. this allows to remove ext4.allow_unsupported=1

this flag is only available in sle12 and sle15, and since those
default to ext3 anyway there is no need to set it. this also
fixes ext4 for all non-suse distros.