Batch calling external commands in a loop
In many cases rather than invoking the same command in a loop
we can collect all the arguments and use xargs to batch up the
invocations. this reduces the per-loop iteration overhead.

In the smallest package in tumbleweed this saves about 1s
of execution overhead on a very fast x86_64 cpu, and 3-5s on
a slower aarch64 cpu. larger packages that install more dependencies
have better outcomes.

measurement of just init_buildsystem for xterm package setup on
a tmpfs store:

before:

  2.63user 1.02system 0:03.32elapsed 109%CPU (0avgtext+0avgdata 36232maxresident)k
  584inputs+496288outputs (9major+576786minor)pagefaults 0swaps

after:

  1.81user 0.74system 0:02.26elapsed 112%CPU (0avgtext+0avgdata 36052maxresident)k
  584inputs+496288outputs (20major+246016minor)pagefaults 0swaps