Move virtio thread enabling into the fixup
The position before this commit was not correct, as the code
afterwards changes kvm_device to non-virtio if no special
initrd is found. It gets changed back to virtio in the fixup
call later on if a special kernel is used.

So we need to do the thread option adding after the fixup code
changed the device back to virtio.