deb: defer dpkg triggers until all packages are installed, and disable man-db altogether (#813)
* deb: do not run dpkg triggers on each package installation

They are designed to be ran at the end of the installation of a batch,
so if supported pass --no-triggers. The existing dpkg --configure
call will then run the triggers at the end of the installation phase,
saving some time.

* deb: preseed man-db debconf to avoid running db update

We don't need to build the database of installed manpages when building
packages. Create/append to the debconf database (so that we don't need
to run the debconf tool, which might not be available until later) to
ensure man-db will skip building/updating the database.