Add Changelog patching when building DSC format 3.0. (#831)
'debtransform' is patching 'changelog' when make *.diff for DSC 1.0.
It adds a 'changelog' entry with the actual build package version.

Missing the actual build package version in 'changelog' is the reason for misbehavior because after 'debtransform' finishes its work, the version from the last 'changelog' entry is used for determining the package version.

This commit contains changes to maintain 'changelog' patching during the build DSC 3.0 in the same way as for DSC 1.0.
To do it, the patching procedure was separated into a separate function dopatch_changelog then used for both building scenarios.

In case of building DSC 3.0 added 'changelog' catching in two different cases:
- When adding a file with the name "debian.changelog".
- When adding tar contents containing "debian/changelog".