Add handling of non-compressed tar when creating Debian archive for DSC 3.0
Before these changes, the function that creates the Debian archive
`dotar_quilt()`unpacked compressed archives like `gz`, `xz`, `bz2` to
a `temp` folder and used those uncompressed `tar`s for concatenation
to the final archive. Then the function was removing the intermediate
file at the end.

This patch adds possibility to pass to the `dotar_quilt()` function
an uncompressed `tar` archive. The function skips the copying of
the uncompressed `tar` to the `temp` folder and uses the original path
for concatenation until a `changelog` in the uncompressed  `tar` found.