INCOMPATIBLE CHANGE: get rid off the power8 cpu limitation
Old build script code was limiting cpu to power8 when building on
power9 always.

As this blocks any future development and also is only done on this
particular platform this way we change it by default.
So KVM on power9 is not special anymore and uses the cpu by default
as with any other architecture or VM type.

A buildflag kvm-max-cpu-compat-power8 can be set to restore the old
behaviour.