Resubmit: Add handling of non-compressed tar when creating Debian archive for DSC 3.0 (#862)
* Add handling of non-compressed tar when creating Debian archive for DSC 3.0

Before these changes, the function that creates the Debian archive
`dotar_quilt()`unpacked compressed archives like `gz`, `xz`, `bz2` to
a `temp` folder and used those uncompressed `tar`s for concatenation
to the final archive. Then the function was removing the intermediate
file at the end.

This patch adds possibility to pass to the `dotar_quilt()` function
an uncompressed `tar` archive. The function skips the copying of
the uncompressed `tar` to the `temp` folder and uses the original path
for concatenation until a `changelog` in the uncompressed  `tar` found.

* Added suggested changes from review

- System call with argument list instead of one string. (For piped call IPC::Run used)
- Copy non-compressed tar instead of handled by code logic
- Clear code from unnecessary quotes and brackets

* Avoid IPC::Run usage