Wrap chroot + su in a function to consistently set up environment for builds
Since "su -" cleans the environment, we need to tell it to keep
variables we want to preserve. To make this easier, create a new wrapper
su_deb_chroot that runs chroot and su and sets up the environment.

Since DEBIAN_* variables are only needed when preparing the build
chroot, don’t set them in su_deb_chroot.

Signed-off-by: Andrej Shadura <andrew.shadura@collabora.co.uk>