changelog2spec: fix time zone handling
The new --timestampfile option writes the time stamp of the
first changes entry to the specified file. We wrote the wrong
time because we did the time zone compensation while parsing
the entry.

The commit moves the time zone compensation to the correct
place, i.e. when converting the time back to a string. Also,
we skip the compensation when the --fulltimestamps option
is given, as we always write UTC time stamps in that case.

(The purpose of the compensation is to use the time zone of
the first changes entry for all the following entry. It is
somewhat historic and we might drop it altogether in the future.)