Adding debian support to generate_sbom
Added support for debian images to generate_sbom.
It checks if it is a debian based image by evaluating the os-release file.

It introduces some new BuildFlags to configure "special" behaviour:

`spdx-declared-license:NOASSERTION` licenseDeclared in spdx info for a package is set to "NOASSERATION",
if not set the found license from the package information is used (like for "licenseConcluded").

The license information is taken from the package's copyrightfile.

The generation of spdx' file section can be disabled, if just the package information is wanted.
This can be switched by buildflag "spdx-files-generation". To switch it off it has to be set to "no":
`spdx-files-generation:no'. Default is to generate also the files information.

As debian packages does not have a vendor information this information is taken from the
OBS vendor macro/OBS instance name.