vm-build: ext3 & ext4: fix disk space allocation
We did fallocate the root device explicit to ensure available disk space
according to the config. However, the mkfs is operation with discard and
frees the disk space again.

This leads to over-commited when having multiple worker instances.

So we explicit disable the punch hole/fallocate call inside of mkfs.ext3
and mkfs.ext4 by setting nodiscard.

An obs worker instance will disable itself when running into this situation.