pbuild: Use the Release value from _config
For reproducible-builds I would like
to be able to produce identical binaries to OBS.

In prjconf I added `Release: 1.1`
but without this patch, pbuild ignored this.

We substitute the OBS CI_CNT and B_CNT counters with 1 to remain compatible.

Because there is a default value of `<CI_CNT>.<B_CNT>`,
this will change the default package names from foo-123-0 to foo-123-1.1