mkosi: use obscpio directory as source if present
When using the obs_scm service the content will be unpacked in
SOURCES/foobar/ but mkosi will run from SOURCES/ and not find the recipes.
If SOURCES/foobar.obsinfo is present and SOURCES/foobar/mkosi.conf is also
present, then use SOURCES/foobar/ as the source dir, so that we can
store recipes in git and use obs_scm.