mkosi: Pass --checksum=yes instead of --checksum (#1059)
In https://github.com/systemd/mkosi/pull/3507 we've made a slight
change to the command line parser to remove ambiguity. This requires
all boolean options to receive an explicit argument now. This commit
changes OBS to pass all boolean options with an argument. This was
already supported by mkosi and will still work even with older
versions of mkosi so the change should be safe to make.