Implement new "partial copyin/out" vm feature
This does not do a full copyout from the vm image, but instead
the new files are put in the image via a tar file that is
injected in the system via debugfs.

We do this because a copyout/copyin of the complete image
clobbers all file owners and devices.