Allow to transport jobs/smp_mflags over the RPM_BUILD_NCPUS envvar
This is needed for reproducible builds, as rpm >= 4.20 stores the
expanded spec file in the src rpm.

We do this by setting RPM_BUILD_NCPUS to the jobs parameter before
calling rpmbuild. Also, we only overwrite the jobs/cmp_mflags
macros if they do not use the RPM_BUILD_NCPUS env var.