Support importing preinstallimage as container image
Implement the option to user preinstallimage as base container image for podman and docker
build environments. This speeds up the build process since there would be no need to untar
the preinstallimage and install the packages inside it. It would only import the preinstallimage
if it doesn't already find it implemented in preinstall_container_image().
For finding the package id and HDRMD5 of packages in preinstallimage it prioritizes the preinstallimageinfo
and if not found it would fallback to previous method of reading unpacked preinstallimage.
Adding randomization for container name in both podman and docker to avoid issues in concurrent builds.